requirements.txt
fastapi
uvicorn
tensorflow==2.x.x
Pillow
numpy